POSTing works, uploading and respond, work on seeking
You can now upload a file to the server and get a response, just
would need to hook that into the rest of our code.
1. get file they uploaded
2. do inference on it
3. return with boxes, actor names, and probabilities
4. stick that into HTML

Still need to fix the seeking with the two bars to get the clip
that they want out of the uploaded video. Will need to send those
time codes with the data to the server.